pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "prompt-armor"
version = "0.1.0"
description = "Open-core LLM prompt security analysis — detect prompt injections, jailbreaks, and other attacks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "prompt-armor contributors" }]
keywords = ["llm", "security", "prompt-injection", "jailbreak", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
    "pydantic>=2.0",
    "numpy>=1.24",
]

[project.optional-dependencies]
ml = [
    "onnxruntime>=1.16",
    "sentence-transformers>=2.2",
    "faiss-cpu>=1.7",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.8",
    "types-PyYAML>=6.0",
]
mcp = [
    "mcp>=1.0",
]
all = [
    "prompt-armor[ml,dev,mcp]",
]

[project.scripts]
prompt-armor = "prompt_armor.cli.main:cli"
prompt-armor-mcp = "prompt_armor.mcp.server:main"

[project.urls]
Homepage = "https://github.com/prompt-armor/prompt-armor"
Repository = "https://github.com/prompt-armor/prompt-armor"
Issues = "https://github.com/prompt-armor/prompt-armor/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/prompt_armor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.ruff]
target-version = "py310"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501", "E402"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
